// src/events/types/BodyStyles.ts
export enum BodyStyles {
  
  // bus body style
  Bus = 'Bus',

  // convertible style
  Convertible = 'Convertible',

  // Coupe style
  Coupe = 'Coupe',

  // Crossover style
   Crossover = 'Crossover',

   // hatchback style
   hatchback = 'hatchback',

   // Minivan style
   Minivan = 'Minivan',

   // Pickup style
   Pickup = 'Pickup',

   // Sedan style
   Sedan = 'Sedan',

   // Other style
   Other = 'Other',

   // Station Wagon style
   StationWagon = 'Station Wagon'
}